Wake up integration test suite
