+++ src/extension-test/integration/suite/util.ts
@@ -1,0 +1,3 @@
+import * as path from 'path';
+
+export const testDataDir = path.join(__dirname, ...["..", "..", "..", ".."], 'test-data', 'integration-test');
